Getting remote tasks working (no celery yet)
Works on #66, #48
Got tasks/remote and remote/h2oline taking care of all the h2oline samples, and moved add_new_sample to remote/__init__.py
Usings parsedatetime to deal with the natual language in the strings.
Added a h2oline sensor with CFS readings, and one with lake height.